feat(repo_identity): relocate pure helpers from nexus.registry (RDR-137 Phase 5.2b, nexus-tts0d.21)
Moves 6 pure helpers from nexus.registry into a new
nexus.repo_identity module:
- _resolve_main_repo
- _repo_identity (2-tuple)
- _repo_identity_with_main (3-tuple)
- _safe_collection
- _sanitise_owner_segment
- _resolve_repo_collection
- list_sibling_collections

nexus.registry now imports them all back so the ~15 legacy import
sites continue to work for one release cycle (back-compat shim).
RepoRegistry class stays in nexus.registry pending nexus-tts0d.20.

The helpers were never registry-coupled — they implement
git-worktree-stable repo identity (`git rev-parse --git-common-dir`)
and the RDR-103 conformant collection-naming rules (4-segment shape,
chroma-compatible character sanitisation). Both outlive the registry.

Closes nexus-tts0d.21.
